Fianl Version
All cards are initially face down. There are two difficulty options: Easy and Hard. The time is counting down. The number of times each card is turned over is limited, and once the number is exhausted, it will fail and be recorded as 0 points. Hard difficulty will have more time and cheats. After completing the game, a score will be given according to the overall performance and accumulated on the leaderboard on the right.